pyproject.toml
[project]
name = "semantic-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "albumentations>=2.0.5",
    "numpy>=2.0.2",
    "opencv-python>=4.11.0.86",
    "pytorch-lightning>=2.5.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "segmentation-models-pytorch>=0.4.0",
    "torch>=2.6.0",
    "torchmetrics>=1.7.0",
    "torchvision>=0.21.0",
]
